requirement.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
langchain_community
requests
python-dotenv
duckduckgo-search
google-api-python-client
